feat: chat editor
